+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Driver plugins: Result::fetch_field() reports the type name in native_type instead of the MySQL number in type and charsetnr
 - Method manifest() to customize or disable the web app manifest
 - Method verifyLoginToken() to accept the login form of another website
+- Method explain() to customize EXPLAIN in SQL command
 ### Internal
 - Test Oracle
 - Distribute the tests in the release archive (bug #1332)